Init 4.1
Had a problem with React showing a blank screen so Steffen and I tried to troubleshoot a lot of solutions on why it did not work. Made some changes on some methods that I made. Turns out it was that I was trying to load an object instead of a string that I forgot to finish.